+++ addon/locales/en/components/flexberry-layers-attributes-panel.js
@@ -3,6 +3,7 @@ export default {
   'select-all': 'Select all',
   'select-feature': 'Find an item on the map',
   'clear-found': 'Clear the found items',
+  'find-intersection': 'Find intersecting polygons',
   'delete-selected': 'Delete the selected items',
   'select-item': 'Select',
   'edit-item': 'Edit',